+++ lib/wechat_pay/api/query_order.ex
@@ -15,7 +15,7 @@ defmodule WechatPay.API.QueryOrder do
       iex> WechatPay.API.QueryOrder.request(%{out_trade_no: "1415757673"})
       {:ok, data}
   """
-  @spec request(map) :: {:ok, map} | {:error, any}
+  @spec request(map) :: {:ok, map} | {:error, WechatPay.Error.t | HTTPoison.Error.t}
   def request(params \\ %{}) do
     order =
       WechatPay.API.QueryOrder.Order