+++ lib/wechat_pay/api/query_refund.ex
@@ -19,7 +19,7 @@ defmodule WechatPay.API.QueryRefund do
       iex> WechatPay.API.QueryRefund.request(params)
       {:ok, data}
   """
-  @spec request(map) :: {:ok, map} | {:error, any}
+  @spec request(map) :: {:ok, map} | {:error, WechatPay.Error.t | HTTPoison.Error.t}
   def request(params \\ %{}) do
     request_data =
       WechatPay.API.QueryRefund.RequestData